+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   CLI for generating node configs and starting nodes has been extended
   in order to reflect these changes. (#1222, #1096, #1235)
 
+- Changed a response for `/healthcheck` endpoint. (#1252)
+
+- Removed a sub-command `generate-testnet` from CLI. (#1264)
+
 #### exonum-crypto
 
 - Renamed `create_keys_file` function to `generate_keys_file`